django/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
requests==2.32.3
dj-database-url==3.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
whitenoise==6.8.2
python-decouple==3.8
django-debug-toolbar==4.4.6
django-extensions==3.2.3
Pillow==11.0.0
